move timepoint sorting code to separate file
