+++ test/schedule/data_test.exs
@@ -578,109 +578,6 @@ defmodule Schedule.DataTest do
     end
   end
 
-  describe "timepoints_for_route/4" do
-    test "returns all timepoint IDs for this route (either direction), sorted" do
-      route_patterns = [
-        %RoutePattern{
-          id: "rp1",
-          name: "rp1",
-          route_id: "r1",
-          direction_id: 0,
-          representative_trip_id: "t1"
-        },
-        %RoutePattern{
-          id: "rp2",
-          name: "rp2",
-          route_id: "r2",
-          direction_id: 0,
-          representative_trip_id: "t2"
-        },
-        %RoutePattern{
-          id: "rp3",
-          name: "rp3",
-          route_id: "r1",
-          direction_id: 1,
-          representative_trip_id: "t3"
-        }
-      ]
-
-      stop_times_by_id = %{
-        "t1" => [
-          %StopTime{stop_id: "s1", time: 1, timepoint_id: "tp1"},
-          %StopTime{stop_id: "s7", time: 2, timepoint_id: nil}
-        ],
-        "t2" => [
-          %StopTime{stop_id: "s2", time: 1, timepoint_id: "tp2"},
-          %StopTime{stop_id: "s3", time: 2, timepoint_id: "tp3"}
-        ],
-        "t3" => [
-          %StopTime{stop_id: "s4", time: 1, timepoint_id: "tp4"},
-          %StopTime{stop_id: "s5", time: 2, timepoint_id: "tp1"}
-        ]
-      }
-
-      timepoints_by_id = %{
-        "tp1" => %Timepoint{id: "tp1", name: "tp1 name"},
-        "tp2" => %Timepoint{id: "tp2", name: "tp2 name"},
-        "tp3" => %Timepoint{id: "tp3", name: "tp3 name"},
-        "tp4" => %Timepoint{id: "tp4", name: "tp4 name"}
-      }
-
-      assert Data.timepoints_for_route(route_patterns, "r1", stop_times_by_id, timepoints_by_id) ==
-               [
-                 %Timepoint{id: "tp4", name: "tp4 name"},
-                 %Timepoint{id: "tp1", name: "tp1 name"}
-               ]
-    end
-
-    test "groups timepoints together even when they're on different stops" do
-      route_patterns = [
-        %RoutePattern{
-          id: "rp1",
-          name: "rp1",
-          route_id: "r1",
-          direction_id: 1,
-          representative_trip_id: "t1"
-        },
-        %RoutePattern{
-          id: "rp2",
-          name: "rp2",
-          route_id: "r1",
-          direction_id: 1,
-          representative_trip_id: "t2"
-        }
-      ]
-
-      stop_times_by_id = %{
-        "t1" => [
-          %StopTime{stop_id: "s1", time: 1, timepoint_id: "tp1"},
-          %StopTime{stop_id: "s3b", time: 2, timepoint_id: "tp3"}
-        ],
-        "t2" => [
-          %StopTime{stop_id: "s1", time: 1, timepoint_id: "tp1"},
-          %StopTime{stop_id: "s2", time: 2, timepoint_id: "tp2"},
-          %StopTime{stop_id: "s3a", time: 3, timepoint_id: "tp3"},
-          %StopTime{stop_id: "s4", time: 4, timepoint_id: "tp4"}
-        ]
-      }
-
-      timepoints_by_id = %{
-        "tp1" => %Timepoint{id: "tp1", name: "tp1 name"},
-        "tp2" => %Timepoint{id: "tp2", name: "tp2 name"},
-        "tp3" => %Timepoint{id: "tp3", name: "tp3 name"},
-        "tp4" => %Timepoint{id: "tp4", name: "tp4 name"}
-      }
-
-      assert Data.timepoints_for_route(route_patterns, "r1", stop_times_by_id, timepoints_by_id) ==
-               [
-                 %Timepoint{id: "tp1", name: "tp1 name"},
-                 %Timepoint{id: "tp2", name: "tp2 name"},
-                 %Timepoint{id: "tp3", name: "tp3 name"},
-                 %Timepoint{id: "tp4", name: "tp4 name"}
-               ]
-    end
-  end
-
   describe "minischedule_run" do
     test "returns run for the trip" do
       trip = %Trip{
